% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_df.R
\name{quantile_df}
\alias{quantile_df}
\title{Create tibble of quantiles}
\usage{
quantile_df(x, p = c(0.25, 0.5, 0.75))
}
\arguments{
\item{x}{numeric vector whose quantiles are wanted, or column name}

\item{p}{probability values}
}
\value{
tibble
}
\description{
Creates a single-row tibble with quantiles and pr values.
}
