% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowAny.R
\name{rowAny}
\alias{rowAny}
\title{Helper for filtering on any logical in a row}
\usage{
rowAny(x)
}
\arguments{
\item{x}{logical}
}
\value{
logical
}
\description{
Helper function for filtering on any TRUE in a row; intended as a replacement
for the depreciated \code{any_vars()} syntax.
}
\examples{
library(tidyverse)
library(rzdr)
diamonds \%>\%
  filter(rowAny(across(c(x, y, z), ~ .x > 0)))
}
